% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_pfun.R
\name{add_bayes_forest}
\alias{add_bayes_forest}
\title{Add Bayesmeta diamond to a forest plot}
\usage{
add_bayes_forest(
  p,
  bm,
  color = "black",
  label = "Bayesmeta",
  mu_estimate = "median"
)
}
\description{
Internal helper function used by \code{\link[=autoplot.confMeta]{autoplot.confMeta()}} to add the
Bayesian diamond corresponding to a \code{bayesmeta} object at the bottom
of an existing forest plot.
}
\keyword{internal}
