% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate_tau2}
\alias{estimate_tau2}
\alias{estimate_phi}
\title{Estimation of between-study heterogeneity}
\usage{
estimate_tau2(estimates, SEs, ...)

estimate_phi(estimates, SEs)
}
\arguments{
\item{estimates}{Numeric vector of effect estimates.}

\item{SEs}{Numeric vector containing the standard errors of the effect
estimates.}

\item{...}{Further arguments that are passed to \code{\link[meta]{metagen}}.}
}
\value{
For \code{estimate_tau2}: the estimated heterogeneity parameter
\eqn{\tau^2}, a \code{numeric} vector of length 1.

For \code{estimate_phi}: the estimated heterogeneity parameter
\eqn{\phi}, a \code{numeric} vector of length 1.
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{\code{estimate_tau2} estimates the between-study heterogeneity
\eqn{\tau^2} using an additive model. The resulting parameter
\eqn{\tau^2} is estimated through a call to
\code{\link[meta]{metagen}} with \code{TE = estimates} and
\code{seTE = SEs}. Other arguments to \code{\link[meta]{metagen}} can be
passed via the \code{...} argument. If no arguments are passed via
\code{...}, the following defaults are applied.
\\itemize\{
  \\item\{\code{sm} = \code{"md"}\}
  \\item\{\code{method.tau} = \code{"REML"}\}
  \\item\{\code{control} = \code{list(maxiter = 1e5, stepadj = 0.25)}\}
\}
However, each of these defaults can be overwritten via \code{...}.
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{\code{estimate_phi} estimates the the between-study heterogeneity
\eqn{\phi} using a multiplicative model. The function is a modified
version of Page 3 of: \\cr \\cr
Accounting for heterogeneity in meta-analysis using a
multiplicative model - an empirical study,
Mawdsley D. etal. 2016. DOI: 10.1002/jrsm.1216 \\cr \\cr

However, in contrast to the paper, it does not truncate the MSE at 1.
\\cr \\cr
}\if{html}{\out{</div>}}
}
\note{
In the paper the weights should be squared!?
}
\examples{
    # Example estimates and std. errors
    estimates <- c(0.21, 0.53, 0.24, 0.32, 0.19)
    SEs <- c(0.19, 0.39, 0.7, 1, 0.97)

    # Estimating heterogeneity using the multiplicative model
    estimate_tau2(estimates = estimates, SEs = SEs)
    # Estimating heterogeneity using the multiplicative model
    estimate_phi(estimates = estimates, SEs = SEs)
}
