% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimatePhi}
\alias{estimatePhi}
\title{Estimate phi for multiplicative heterogeneity}
\usage{
estimatePhi(thetahat, se)
}
\arguments{
\item{thetahat}{Numeric vector of parameter estimates.}

\item{se}{Numeric vector of standard errors.}
}
\value{
Estimated phi, a \code{numeric} vector of length 1.
}
\description{
A modified version of Page 3 of: \cr \cr
Accounting for heterogeneity in meta-analysis using a multiplicative model -
an empirical study, Mawdsley D. etal. 2016. DOI: 10.1002/jrsm.1216

This function, in contrast to the paper, does not truncate the MSE at 1.
}
\note{
In the paper the weights should be squared!?
}
\examples{
thetahat <- c(0.21, 0.53, 0.24, 0.32, 0.19)
se <- c(0.19, 0.39, 0.7, 1, 0.97)
estimatePhi(thetahat = thetahat, se = se)|

thetahat[1] <- 7
estimatePhi(thetahat = thetahat, se = se)
}
\keyword{internal}
