% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest2ci.R
\name{forest2CI}
\alias{forest2CI}
\title{Extract the confidence interval from forest plots}
\usage{
forest2CI(forest)
}
\arguments{
\item{forest}{The output of a call to \code{\link[hMean]{ForestPlot}}}
}
\value{
A list with confidence intervals from all the methods specified
in the call to \code{ForestPlot}.
}
\description{
Extract the confidence interval from forest plots
}
\examples{
thetahat <- c(-0.49, -0.17, -0.52, -0.48, -0.26, -0.36, -0.47, -0.3, -0.15,
              -0.28)
se <- c(0.346945150708765, 0.214289651908355, 0.239800324754587,
        0.372455823554997, 0.25000459389308, 0.295923805016299,
        0.219391786477601, 0.14796190250815, 0.270413132170067,
        0.500009187786161)
pValueFUN <- c("hMean", "k-Trials", "Pearson", "Edgington", "Fisher")
distr <- c("chisq")
heterogeneity <- "none"
forest <- ForestPlot(
    thetahat = thetahat,
    se = se,
    distr = distr,
    pValueFUN = pValueFUN,
    heterogeneity = heterogeneity
)
forest2CI(forest)

}
