% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestPlot.R
\name{ForestPlot}
\alias{ForestPlot}
\title{Plot the forest plot for a given meta-analysis}
\usage{
ForestPlot(
  thetahat,
  se,
  level = 0.95,
  distr = c("chisq", "f"),
  pValueFUN = c("hMean", "k-Trials", "Pearson", "Edgington", "Fisher"),
  heterogeneity = c("none", "additive", "multiplicative"),
  diamond_height = 0.5,
  v_space = 1.5,
  studyNames = NULL,
  xlim = NULL,
  show_studies = TRUE,
  scale_diamonds = FALSE
)
}
\arguments{
\item{thetahat}{Numeric vector of parameter estimates.}

\item{se}{Numeric vector of standard errors.}

\item{level}{Numeric vector of length 1 specifying the level of the
confidence interval. Defaults to 0.95.}

\item{distr}{Character vector of length 1 or 2. Valid options are
either \code{"f"} or \code{"chisq"} or both. Denotes the distribution
used in cases where \code{pValueFUN = "hMean"} but ignored otherwise.}

\item{pValueFUN}{Character vector of length 1, 2 or 3. Valid options consist
of any combination of \code{"hMean"}, \code{"k-Trials"}, or \code{"Pearson"}.
\code{"hMean"} will add a line using \code{\link[confMeta]{hMeanChiSqMu}} as
the p-value function, \code{"k-Trials"} adds a line using
\code{\link[confMeta]{kTRMu}} as the p-value function, and \code{"Pearson"}
uses \code{\link[confMeta]{pPearsonMu}} as the p-value function.}

\item{heterogeneity}{The heterogeneity model used to adjust the standard
errors \code{se}. Valid options are any combination of \code{"none"},
\code{"additive"}, or \code{"multiplicative"}. See also
\code{\link[confMeta]{hMeanChiSqMu}}.}

\item{diamond_height}{The height of the polygons. Must be a numeric vector of
length 1. Defaults to 0.5.}

\item{v_space}{A numeric vector of length 1 indicating the vertical space
between the different rows of the plot. Default is 1.5.}

\item{studyNames}{Either \code{NULL} (the default) or a character vector of
the same length as \code{thetahat} indicating the names of the individual
studies.}

\item{xlim}{Either \code{NULL} (default) or a numeric vector of length 2,
indicating the limits of the x-axis.}

\item{show_studies}{Must be either \code{TRUE} (default) or \code{FALSE}.
If \code{TRUE}, the confidence intervals for the individual studies will
be shown in the forest plot. Otherwise, the plot will only show the diamonds
for the meta-analyses.}

\item{scale_diamonds}{Must be either \code{TRUE} or \code{FALSE} with
\code{FALSE} being the default. If \code{TRUE}, the maximum height of the
diamond is always \code{diamond_height}, even if the p-value function never
reaches 1. If \code{FALSE}, the the maximum height of the diamond is
\code{diamond_heigth} * max(p-value function).}
}
\value{
The function invisibly returns a list with the following elements:
\describe{
\item{plot}{An object of class \code{ggplot}. It shows a forest plot for
the individual studies as well as the overall estimate derived
using the specified methods.}
\item{p_0}{A data frame containing the p-values for the null-effect
according to the specified methods.}
}
}
\description{
Plot the forest plot for a given meta-analysis
}
\examples{
thetahat <- c(-0.49, -0.17, -0.52, -0.48, -0.26, -0.36, -0.47, -0.3, -0.15,
              -0.28)
se <- c(0.346945150708765, 0.214289651908355, 0.239800324754587,
        0.372455823554997, 0.25000459389308, 0.295923805016299,
        0.219391786477601, 0.14796190250815, 0.270413132170067,
        0.500009187786161)
pValueFUN <- c("hMean", "k-Trials", "Pearson", "Edgington", "Fisher")
distr <- c("chisq")
heterogeneity <- "none"
ForestPlot(
    thetahat = thetahat,
    se = se,
    distr = distr,
    pValueFUN = pValueFUN,
    heterogeneity = heterogeneity
)

}
