% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPvalueFunction.R
\name{ggPvalueFunction}
\alias{ggPvalueFunction}
\title{Plot the p-value function(s) for a given set of studies}
\usage{
ggPvalueFunction(
  thetahat,
  se,
  level = 0.95,
  distr = c("f", "chisq"),
  pValueFUN = c("hMean", "k-Trials", "Pearson"),
  heterogeneity = c("none", "additive", "multiplicative"),
  xlim = c(min(thetahat - 3 * se), max(thetahat + 3 * se))
)
}
\arguments{
\item{thetahat}{Numeric vector of parameter estimates.}

\item{se}{Numeric vector of standard errors.}

\item{level}{Numeric vector of length 1 specifying the level of the
confidence interval. Defaults to 0.95.}

\item{distr}{Character vector of length 1 or 2. Valid options are
either \code{"f"} or \code{"chisq"} or both. Denotes the distribution
used in cases where \code{pValueFUN = "hMean"} but ignored otherwise.}

\item{pValueFUN}{Character vector of length 1 or 2. Valid options are
either \code{"hMean"} or \code{"k-Trials"} or both. \code{"hMean"} will
add a line using \code{\link[hMean]{hMeanChiSqMu}} as the p-value
function whereas \code{"k-Trials"} adds a line using
\code{\link[hMean]{kTRMu}} as the p-value function.}

\item{heterogeneity}{One or more of c("none", "additive", "multiplicative").
If heterogeneity = "none" p-values are returned for the passed se without
any adaption. If heterogeneity = "additive", the standard errors se are
reassigned the value of sqrt(se^2 + tau2) before computation of the p-values.
If heterogeneity = "multiplicative", the standard errors se are multiplied
with the value of phi before computation of the p-values.
Defaults to "none".}

\item{xlim}{Numeric vector of length 2 denoting the range of values
on the x-axis, i.e. the effect size.}
}
\description{
Plot the p-value function(s) for a given set of studies
}
