% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hMeanChiSqMu.R
\name{hMeanChiSqMu}
\alias{hMeanChiSqMu}
\title{Calculate the p-value}
\usage{
hMeanChiSqMu(
  thetahat,
  se,
  w = rep(1, length(thetahat)),
  mu = 0,
  phi = estimatePhi(thetahat, se),
  tau2 = estimateTau2(thetahat, se),
  alternative = "none",
  distr = c("chisq", "f"),
  heterogeneity = c("additive", "multiplicative"),
  bound = FALSE
)
}
\arguments{
\item{thetahat}{Numeric vector of parameter estimates.}

\item{se}{Numeric vector of standard errors.}

\item{w}{Numeric vector of weights.}

\item{mu}{A numeric vector containing null hypothesis value(s). Defaults to 0.}

\item{phi}{A numeric vector of length 1. TODO: list all other restrictions on the parameter.}

\item{tau2}{A numeric vector of length 1. TODO: list all other restrictions on the parameter.}

\item{alternative}{Either \code{"greater"}, \code{"less"}, \code{"two.sided"}, or \code{"none"} (default).
Specifies the alternative to be considered in the computation of the p-value.}

\item{distr}{The distribution to use for the calculation of the p-value. Currently, the options are
\code{"f"} (F-distribution) and \code{"chisq"} (Chi-squared distribution).}

\item{heterogeneity}{The method used to estimate heterogeneity. Implemented options are \code{"additive"} and \code{"multiplicative"}.}

\item{bound}{If \code{FALSE} (default), p-values that cannot be computed are reported as \code{NaN}.
If \code{TRUE}, they are reported as "> bound".}
}
\value{
Returns the p-value from the harmonic mean chi-squared test
based on study-specific estimates and standard errors.
}
\description{
Calculate the p-value
}
