% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kTRMu.R
\name{kTRMu}
\alias{kTRMu}
\title{Implementation of the "k-Trials" rule.}
\usage{
kTRMu(
  thetahat,
  se,
  mu,
  phi = NULL,
  tau2 = NULL,
  alternative = "none",
  heterogeneity = c("additive", "multiplicative")
)
}
\arguments{
\item{thetahat}{Numeric vector of parameter estimates.}

\item{se}{Numeric vector of standard errors.}

\item{mu}{A numeric vector containing null hypothesis value(s).}

\item{phi}{A numeric vector of length 1. TODO: list all other restrictions on the parameter.}

\item{tau2}{A numeric vector of length 1. TODO: list all other restrictions on the parameter.}

\item{alternative}{Either \code{"greater"}, \code{"less"}, \code{"two.sided"}, or \code{"none"} (default).
Specifies the alternative to be considered in the computation of the p-value.}

\item{heterogeneity}{The method used to estimate heterogeneity. Implemented options are \code{"additive"} and \code{"multiplicative"}.}
}
\value{
The corresponding p-value given mu under the null-hypothesis
}
\description{
Implementation of the "k-Trials" rule.
}
\examples{
thetahat <- c(0.38041824, -0.22170681, -0.09155859)
se <- c(0.2256202, 0.2432796, 0.1051743)
phi <- estimatePhi(thetahat = thetahat, se=se)
tau2 <- estimateTau2(thetahat, se, control = list(stepadj = 0.5, maxiter = 1000, threshold = 1e-6))
mymu <- seq(min(thetahat - 3 * se), max(thetahat + 3 * se), length.out=1000)
p_add <- kTRMu(thetahat, se, mymu, tau2 = tau2, phi = phi, heterogeneity = "additive")
p_mult <- kTRMu(thetahat, se, mymu, tau2 = tau2, phi = phi, heterogeneity = "multiplicative")
}
