% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value_functions.R
\name{p_value_functions}
\alias{p_value_functions}
\title{Overview of p-value combination functions}
\description{
The \code{p_value_functions} family is useful to combine
individual study effect estimates and standard errors into a single
meta-analytical \emph{p}-value function. These functions are the
building blocks for the \code{\link{confMeta}} function.
}
\details{
\strong{Vectorization over mu:}
All of the \emph{p}-value functions in this package are vectorized over the
\code{mu} argument. This means that you can pass a sequence of null values
(e.g., \code{seq(-2, 2, length.out = 1000)}) and the function will
evaluate the combined \emph{p}-value for every point. This is specifically
designed to construct \emph{p}-value functions
and invert the tests to find confidence intervals at any desired level.

\strong{Available Methods:}
\itemize{
\item \code{\link{p_edgington}}: Classical Edgington method (Sum of \emph{p}-values).
\item \code{\link{p_edgington_w}}: Weighted Edgington method.
\item \code{\link{p_fisher}}: Fisher's method (Log-product of \emph{p}-values).
\item \code{\link{p_pearson}}: Pearson's method (Log-product of complements).
\item \code{\link{p_tippett}}: Tippett's method (Minimum \emph{p}-value).
\item \code{\link{p_wilkinson}}: Wilkinson's method (Maximum \emph{p}-value).
\item \code{\link{p_stouffer}}: Stouffer's method.
\item \code{\link{p_hmean}}: Harmonic mean method.
}
}
\references{
Held, L, Hofmann, F, Pawel, S. (2025). A comparison of combined \emph{p}-value
functions for meta-analysis. \emph{Research Synthesis Methods}, 16:758-785.
\doi{10.1017/rsm.2025.26}
}
\seealso{
\code{\link{confMeta}}
}
