% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_pfun.R
\name{autoplot.confMeta}
\alias{autoplot.confMeta}
\title{Visualizations of \code{confMeta} objects}
\usage{
\method{autoplot}{confMeta}(
  ...,
  type = c("p", "forest"),
  diamond_height = 0.5,
  v_space = 1.5,
  scale_diamonds = TRUE,
  show_studies = TRUE,
  drapery = TRUE,
  reference_methods = c("re", "hk", "hc"),
  xlim = NULL,
  xlab = NULL,
  xbreaks = 0.5,
  bayesmeta = NULL
)
}
\arguments{
\item{...}{One or more objects of class \code{confMeta}.}

\item{type}{A character vector of length 1 or 2. Indicates what type of
plot should be returned. Accepted is any combination of \code{"p"} and
\code{"forest"}. Defaults to \code{c("p", "forest")}.}

\item{diamond_height}{Numeric vector of length 1. Indicates the maximal
possible height of the diamonds in the forest plot. Defaults to 0.5.
This argument is only relevant if \code{type} contains \code{"forest"} and will
be ignored otherwise.}

\item{v_space}{Numeric vector of length 1. Indicates the vertical space
between two diamonds in the forest plot. Defaults to 1.5.
This argument is only relevant if \code{type} contains \code{"forest"} and will
be ignored otherwise.}

\item{scale_diamonds}{Logical vector of lenght 1. Accepted values are either
\code{TRUE} (default) or \code{FALSE}. If \code{TRUE}, the diamond is rescaled to the
interval [0, 1] in cases where the maximum of the p-value
function is not equal to 1. This argument is only relevant if \code{type}
contains \code{"forest"} and will be ignored otherwise.}

\item{show_studies}{Logical vector of lenght 1. Accepted values are either
\code{TRUE} (default) or \code{FALSE}. If \code{TRUE}, the forest plot shows the
confidence intervals for the individual effect estimates. Otherwise,
the intervals are suppressed. This argument is only relevant if \code{type}
contains \code{"forest"} and will be ignored otherwise.}

\item{drapery}{Either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE}, the individual
study effects are represented as drapery plots. If \code{FALSE} the studies
are represented by a simple vertical line at their effect estimates.}

\item{reference_methods}{A character vector of length 1, 2, 3 or 4.
Specifies which reference meta-analysis methods should be shown in
the plot. Valid options are any subset of
\code{c("fe", "re", "hk", "hc")}, which correspond to:
- \code{"fe"}: fixed-effect meta-analysis
- \code{"re"}: random-effects meta-analysis
- \code{"hk"}: Hartung–Knapp adjustment
- \code{"hc"}: Henmi–Copas adjustment.

\if{html}{\out{<div class="sourceCode">}}\preformatted{**Note:** at most one of `"fe"` or `"re"` can be included at a time.
These two methods are mutually exclusive, as only one baseline
meta-analysis model (fixed or random effects) can be displayed in
a single plot. Defaults to `c("re", "hk", "hc")`.
}\if{html}{\out{</div>}}}

\item{xlim}{Either NULL (default) or a numeric vector of length 2 which
indicates the extent of the x-axis that should be shown.}

\item{xlab}{Either NULL (default) or a character vector of length 1 which
is used as the label for the x-axis.}

\item{xbreaks}{Numeric value of length 1. Defines the spacing of the tick
marks along the x-axis, expressed in the same units as the effect size.
For example, \code{xbreaks = 0.5} will show tick marks at
\verb{..., -1.5, -1.0, -0.5, 0, 0.5, 1.0, 1.5, ...}.
Defaults to \code{0.5}. This argument is used in both the \emph{p}-value function
plot and the forest plot to control the resolution of the x-axis ticks.}

\item{bayesmeta}{Either \code{NULL} (default) or an object of class \code{"bayesmeta"},
typically created using \code{\link[bayesmeta:bayesmeta]{bayesmeta::bayesmeta()}}. When provided, the
posterior median (or mean, depending on the \code{mu_estimate} setting) and
95\% credible interval are displayed as an additional diamond at the
bottom of the forest plot for direct comparison with the frequentist
confidence intervals. This argument has no effect on the \emph{p}-value
function plot.}
}
\value{
An object of class \code{ggplot} containing the specified plot(s).
}
\description{
Plot one or more \code{confMeta} objects. Currently, this function
can create two types of plots, the \emph{p}-value function as well as the
forest plot. This allows to compare different \emph{p}-value functions with
each other.

Optionally, a Bayesian meta-analysis object created with the
\code{\link[bayesmeta:bayesmeta]{bayesmeta::bayesmeta()}} function can be supplied via the \code{bayesmeta} argument.
When provided, its posterior summary is displayed as an additional diamond
at the bottom of the forest plot for comparison with the frequentist methods
}
