% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_pfun.R
\name{autoplot.confMeta}
\alias{autoplot.confMeta}
\title{Visualizations of \code{confMeta} objects}
\usage{
\method{autoplot}{confMeta}(
  ...,
  type = c("p", "forest"),
  diamond_height = 0.5,
  v_space = 1.5,
  scale_diamonds = TRUE,
  show_studies = TRUE,
  drapery = TRUE,
  reference_methods = c("re", "hk", "hc"),
  xlim = NULL,
  xlab = NULL
)
}
\arguments{
\item{...}{One or more objects of class \code{confMeta}.}

\item{type}{A character vector of length 1 or 2. Indicates what type of
plot should be returned. Accepted is any combination of \code{"p"} and
\code{"forest"}. Defaults to \code{c("p", "forest")}.}

\item{diamond_height}{Numeric vector of length 1. Indicates the maximal
possible height of the diamonds in the forest plot. Defaults to 0.5.
This argument is only relevant if \code{type} contains \code{"forest"} and will
be ignored otherwise.}

\item{v_space}{Numeric vector of length 1. Indicates the vertical space
between two diamonds in the forest plot. Defaults to 1.5.
This argument is only relevant if \code{type} contains \code{"forest"} and will
be ignored otherwise.}

\item{scale_diamonds}{Logical vector of lenght 1. Accepted values are either
\code{TRUE} (default) or \code{FALSE}. If \code{TRUE}, the diamond is rescaled to the
interval [0, 1] in cases where the maximum of the p-value
function is not equal to 1. This argument is only relevant if \code{type}
contains \code{"forest"} and will be ignored otherwise.}

\item{show_studies}{Logical vector of lenght 1. Accepted values are either
\code{TRUE} (default) or \code{FALSE}. If \code{TRUE}, the forest plot shows the
confidence intervals for the individual effect estimates. Otherwise,
the intervals are suppressed. This argument is only relevant if \code{type}
contains \code{"forest"} and will be ignored otherwise.}

\item{drapery}{Either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE}, the individual
study effects are represented as drapery plots. If \code{FALSE} the studies
are represented by a simple vertical line at their effect estimates.}

\item{reference_methods}{A character vector of length 1, 2 or 3. Valid
options are any combination of \code{c("fe", "re", "hk", "hc")}. Defaults to
\code{c("fe", "re", "hk", "hc")}.}

\item{xlim}{Either NULL (default) or a numeric vector of length 2 which
indicates the extent of the x-axis that should be shown.}

\item{xlab}{Either NULL (default) or a character vector of length 1 which
is used as the label for the x-axis.}
}
\value{
An object of class \code{ggplot} containing the specified plot(s).
}
\description{
Plot one or more \code{confMeta} objects. Currently, this function
can create two types of plots, the \emph{p}-value function as well as the
forest plot. This allows to compare different \emph{p}-value functions with
each other.
}
