% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPvalueFunction.R
\name{ggPvalueFunction}
\alias{ggPvalueFunction}
\title{Plot the p-value function(s) for a given set of studies}
\usage{
ggPvalueFunction(
  thetahat,
  se,
  level = 0.95,
  distr = c("chisq", "f"),
  heterogeneity = c("none", "additive", "multiplicative"),
  pValueFUN = c("hMean", "k-Trials", "Pearson", "Edgington", "Fisher"),
  pValueFUN_args,
  xlim = c(min(thetahat - 2 * se), max(thetahat + 2 * se))
)
}
\arguments{
\item{thetahat}{Numeric vector of parameter estimates.}

\item{se}{Numeric vector of standard errors.}

\item{level}{Numeric vector of length 1 specifying the level of the
confidence interval. Defaults to 0.95.}

\item{distr}{Character vector of length 1 or 2. Valid options are
either \code{"f"} or \code{"chisq"} or both. Denotes the distribution
used in cases where \code{pValueFUN = "hMean"} but ignored otherwise.}

\item{heterogeneity}{The heterogeneity model used to adjust the standard
errors \code{se}. Valid options are any combination of \code{"none"},
\code{"additive"}, or \code{"multiplicative"}. See also
\code{\link[hMean]{hMeanChiSqMu}}.}

\item{pValueFUN}{Any combination of \code{"hMean"}, \code{"k-Trials"}, and
\code{"Pearson"}. This argument determines which methods are plotted in the
resulting image.}

\item{pValueFUN_args}{A named \code{list} with arguments passed to
\code{pValueFUN}. This list should contain all arguments of \code{pValueFUN}
that do not have a default. Arguments \code{thetahat}, \code{se}, and
\code{mu} are automatically passed to \code{pValueFUN} and must thus not be
included in this list.}

\item{xlim}{Numeric vector of length 2 denoting the range of values
on the x-axis, i.e. the effect size.}
}
\value{
An object of class \code{ggplot}. It contains the plot of the p-value
functions specified by the input arguments.
}
\description{
Plot the p-value function(s) for a given set of studies
}
\examples{
thetahat <- c(-0.49, -0.17, -0.52, -0.48, -0.26, -0.36, -0.47, -0.3, -0.15,
              -0.28)
se <- c(0.346945150708765, 0.214289651908355, 0.239800324754587,
        0.372455823554997, 0.25000459389308, 0.295923805016299,
        0.219391786477601, 0.14796190250815, 0.270413132170067,
        0.500009187786161)
pValueFUN <- c("hMean", "k-Trials", "Pearson", "Edgington")
distr <- c("chisq")
heterogeneity <- "none"
ggPvalueFunction(thetahat = thetahat,
                 se = se,
                 xlim = c(-0.8, 0.1),
                 distr = distr,
                 heterogeneity = heterogeneity,
                 pValueFUN = pValueFUN)
}
