% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hMeanChiSqCI.R
\name{hMeanChiSqCI}
\alias{hMeanChiSqCI}
\title{Calculate confidence intervals based on the harmonic mean chi-squared test}
\usage{
hMeanChiSqCI(
  thetahat,
  se,
  level = 0.95,
  alternative = "none",
  check_inputs = TRUE,
  pValueFUN = hMeanChiSqMu,
  pValueFUN_args
)
}
\arguments{
\item{thetahat}{Numeric vector of parameter estimates.}

\item{se}{Numeric vector of standard errors.}

\item{level}{Numeric vector of length 1 specifying the level of the
confidence interval. Defaults to 0.95.}

\item{alternative}{Either \code{"greater"}, \code{"less"}, \code{"two.sided"}, or \code{"none"} (default).
Specifies the alternative to be considered in the computation of the p-value.}

\item{check_inputs}{Either \code{TRUE} (default) or \code{FALSE}. Indicates whether or not to check the input arguments.
The idea of this argument is that if the function is called a large amount of times in an automated manner as for example
in simulations, performance might be increased by not checking inputs in every single iteration. However, setting the argument
to \code{FALSE} might be dangerous.}

\item{pValueFUN}{A function that calculates the p-value. Must have arguments
\code{thetahat} and \code{se} as these are passed by this function.
Must further have an argument \code{mu} that specifies the null-hypothesis.
Defaults to \code{\link[hMean]{hMeanChiSqMu}}.}

\item{pValueFUN_args}{A named \code{list} with arguments passed to
\code{pValueFUN}. This list should contain all arguments of \code{pValueFUN}
that do not have a default. Arguments \code{thetahat}, \code{se}, and
\code{mu} are automatically passed to \code{pValueFUN} and must thus not be
included in this list.}
}
\value{
Returns a list containing confidence interval(s)
obtained by inverting the harmonic mean chi-squared test based on
study-specific estimates and standard errors. The list contains:
\item{CI}{Confidence interval(s).}\cr\cr
If the \code{alternative} is "none", the list also contains:
\item{gamma}{Local minima of the p-value function between the thetahats.}
\item{gammaMean}{Mean of all gammas.}
\item{gammaHMean}{Harmonic mean of all gammas.}
}
\description{
Calculate confidence intervals based on the harmonic mean chi-squared test
}
\examples{
## Simulate thetahat and se
n <- 15
mean <- 0
sd <- 1.1
shape <- 5
rate <- 5
thetahat <- rnorm(n, mean = mean, sd = sd)
se <- rgamma(n, shape = shape, rate = rate)

## Set heterogeneity
heterogeneity <- "none"
phi <- if (heterogeneity == "multiplicative") {
    estimatePhi(thetahat = thetahat, se = se)
} else {
    NULL
}
tau2 <- if (heterogeneity == "additive") {
    estimateTau2(thetahat = thetahat, se = se)
} else {
    NULL
}

## Set confidence level
conf.level <- 0.95

## Determine the p-value functions
funs <- list(
    "pearson" = hMean::pPearsonMu,
    "hMean" = hMean::hMeanChiSqMu,
    "k-Trials" = hMean::kTRMu,
    "edgington" = hMean::pEdgingtonMu,
    "fisher" = hMean::pFisherMu
)

## Calculate the CIs
cis <- lapply(
    funs,
    function(f) {
        hMeanChiSqCI(
            thetahat = thetahat,
            se = se,
            level = conf.level,
            alternative = "none",
            pValueFUN = f,
            pValueFUN_args = list(
                check_inputs = FALSE,
                heterogeneity = heterogeneity,
                phi = phi,
                tau2 = tau2
            )
        )
    }
)
}
