% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hMeanChiSqMu.R
\name{hMeanChiSqMu}
\alias{hMeanChiSqMu}
\title{Calculate the p-value using the harmonic mean chi-squared test.}
\usage{
hMeanChiSqMu(
  thetahat,
  se,
  mu = 0,
  phi = NULL,
  tau2 = NULL,
  heterogeneity = "none",
  alternative = "none",
  check_inputs = TRUE,
  w = rep(1, length(thetahat)),
  distr = "chisq"
)
}
\arguments{
\item{thetahat}{Numeric vector of parameter estimates.}

\item{se}{Numeric vector of standard errors.}

\item{mu}{A numeric vector containing null hypothesis value(s).}

\item{phi}{A numeric vector of length 1. Must be finite and larger than 0. The square root of the argument is used to scale the standard errors.}

\item{tau2}{A numeric vector of length 1.}

\item{heterogeneity}{One of \code{c("none", "additive", "multiplicative")}. If \code{heterogeneity = "none"} p-values are returned for the
passed \code{se} without any adaption. If \code{heterogeneity = "additive"}, the standard errors \code{se} are reassigned the value of \code{sqrt(se^2 + tau2)} before
computation of the p-values. If \code{heterogeneity = "multiplicative"}, the standard errors \code{se} are multiplied with the value of \code{phi} before
computation of the p-values. Defaults to \code{"none"}.}

\item{alternative}{Either \code{"greater"}, \code{"less"}, \code{"two.sided"}, or \code{"none"} (default).
Specifies the alternative to be considered in the computation of the p-value.}

\item{check_inputs}{Either \code{TRUE} (default) or \code{FALSE}. Indicates whether or not to check the input arguments.
The idea of this argument is that if the function is called a large amount of times in an automated manner as for example
in simulations, performance might be increased by not checking inputs in every single iteration. However, setting the argument
to \code{FALSE} might be dangerous.}

\item{w}{Numeric vector of weights.}

\item{distr}{The distribution to use for the calculation of the p-value. Currently, the options are
\code{"f"} (F-distribution) and \code{"chisq"} (Chi-squared distribution). Defaults to \code{"chisq"}.}
}
\value{
Returns the p-value from the harmonic mean chi-squared test
based on study-specific estimates and standard errors.
}
\description{
Calculate the p-value using the harmonic mean chi-squared test.
}
\details{
The function is vectorized over the argument \code{mu}.
}
