% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfun_hmean.R
\name{p_hmean}
\alias{p_hmean}
\title{Harmonic mean's method}
\usage{
p_hmean(
  estimates,
  SEs,
  mu = 0,
  phi = NULL,
  tau2 = NULL,
  heterogeneity = "none",
  alternative = "none",
  check_inputs = TRUE,
  w = rep(1, length(estimates)),
  distr = "chisq"
)
}
\arguments{
\item{estimates}{Numeric vector of study-level effect estimates.}

\item{SEs}{Numeric vector of corresponding standard errors.}

\item{mu}{Numeric \strong{scalar or vector} of null values for the overall effect
(default: 0).}

\item{phi}{Multiplicative heterogeneity parameter (if applicable).}

\item{tau2}{Additive heterogeneity parameter (if applicable).}

\item{heterogeneity}{Character string: \code{"none"} (default),
\code{"additive"}, or \code{"multiplicative"}. Determines whether
standard errors are adjusted for between-study heterogeneity using
\code{tau2} or \code{phi}.}

\item{alternative}{???}

\item{check_inputs}{Logical (default \code{TRUE}). If \code{TRUE},
perform input validation.}

\item{w}{Numeric vector of weights (default: equal weights).}

\item{distr}{Character string specifying the null distribution:
\code{"chisq"} (default) or \code{"f"}.}
}
\value{
A numeric vector of combined \emph{p}-values corresponding
to each value of \code{mu}.
}
\description{
Combines study-level results using the harmonic mean of squared \eqn{z}-statistics.#'
}
\details{
Explain how it is computed ???
}
\examples{
estimates <- c(0.5, 0.8, 0.3)
SEs <- c(0.1, 0.2, 0.1)
p_hmean(estimates, SEs, mu = 0, distr = "f")

}
\references{
???

Held L, Hofmann F, Pawel S. A comparison of combined p-value functions for meta-analysis. \emph{Research Synthesis Methods}, 16:758-785, 2025.
}
\seealso{
Other p-value combination functions: 
\code{\link{p_edgington}()},
\code{\link{p_edgington_w}()},
\code{\link{p_fisher}()},
\code{\link{p_pearson}()},
\code{\link{p_stouffer}()},
\code{\link{p_tippett}()},
\code{\link{p_wilkinson}()}
}
\concept{p-value combination functions}
