% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfun_hmean.R
\name{p_hmean}
\alias{p_hmean}
\title{Harmonic mean method}
\usage{
p_hmean(
  estimates,
  SEs,
  mu = 0,
  phi = NULL,
  tau2 = NULL,
  heterogeneity = "none",
  check_inputs = TRUE,
  w = rep(1, length(estimates)),
  distr = "chisq"
)
}
\arguments{
\item{estimates}{Numeric vector of study-level effect estimates.}

\item{SEs}{Numeric vector of corresponding standard errors.}

\item{mu}{Numeric scalar or vector of null values for the overall effect
(default: 0).}

\item{phi}{Multiplicative heterogeneity parameter (if applicable).}

\item{tau2}{Additive heterogeneity parameter (if applicable).}

\item{heterogeneity}{Character string: \code{"none"} (default),
\code{"additive"}, or \code{"multiplicative"}. Determines whether
standard errors are adjusted for between-study heterogeneity using
\code{tau2} or \code{phi}.}

\item{check_inputs}{Logical (default \code{TRUE}). If \code{TRUE}, perform
input validation.}

\item{w}{Numeric vector of weights (default: equal weights).}

\item{distr}{Character string specifying the null distribution:
\code{"chisq"} (default) or \code{"f"}.}
}
\value{
A numeric vector of combined \emph{p}-values corresponding
to each value of \code{mu}.
}
\description{
Combines study-level results using the harmonic mean combination method
}
\details{
The harmonic mean statistic for \eqn{k} studies is defined as
\deqn{x^2 = \frac{\sum_{i=1}^k \sqrt{w_i}}{\sum_{i=1}^k w_i/z_i^2},}
where \eqn{z_i} and \eqn{{w_i}} are individual study \emph{z}-values and weights,
respectively. Under the global null hypothesis, each \eqn{z_i} is assumed to
follow a standard normal distribution. The harmonic mean statistic then
follows a chi-squared distribution with one degree of freedom. The combined
\emph{p}-value is calculated as the probability of observing a value equal or
greater than \eqn{x^2} from this distribution: \deqn{p_H = \Pr(\chi^2_{1} >
x^2)}
}
\examples{
estimates <- c(0.5, 0.8, 0.3)
SEs <- c(0.1, 0.2, 0.1)
p_hmean(estimates, SEs, mu = 0, distr = "f")

}
\references{
Held, L. (2020). The harmonic mean chi-squared test to substantiate
scientific findings. \emph{Journal of the Royal Statistical Society: Series C
(Applied Statistics)}, 69:697-708. \doi{10.1111/rssc.12410}

Held, L, Hofmann, F, Pawel, S. (2025). A comparison of combined \emph{p}-value
functions for meta-analysis. \emph{Research Synthesis Methods}, 16:758-785.
\doi{10.1017/rsm.2025.26}
}
\seealso{
Other p-value combination functions: 
\code{\link{p_edgington}()},
\code{\link{p_edgington_w}()},
\code{\link{p_fisher}()},
\code{\link{p_pearson}()},
\code{\link{p_stouffer}()},
\code{\link{p_tippett}()},
\code{\link{p_wilkinson}()}
}
\concept{p-value combination functions}
