% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfun_stouffer.R
\name{p_stouffer}
\alias{p_stouffer}
\title{Stouffer's method}
\usage{
p_stouffer(
  estimates,
  SEs,
  mu = 0,
  heterogeneity = "none",
  phi = NULL,
  tau2 = NULL,
  check_inputs = TRUE,
  w = 1/SEs,
  output_p = "two.sided"
)
}
\arguments{
\item{estimates}{Numeric vector of study-level effect estimates.}

\item{SEs}{Numeric vector of corresponding standard errors.}

\item{mu}{Numeric scalar or vector of null values for the overall effect
(default: 0).}

\item{heterogeneity}{Character string: \code{"none"} (default),
\code{"additive"}, or \code{"multiplicative"}. Determines whether
standard errors are adjusted for between-study heterogeneity using
\code{tau2} or \code{phi}.}

\item{phi}{Multiplicative heterogeneity parameter (if applicable).}

\item{tau2}{Additive heterogeneity parameter (if applicable).}

\item{check_inputs}{Logical (default \code{TRUE}). If \code{TRUE},
perform input validation.}

\item{w}{Numeric vector of study weights. Defaults to \code{1/SEs} producing
the same \emph{p}-value as meta-analsis}

\item{output_p}{Character string specifying the combined
\emph{p}-value type: \code{"two.sided"} (default) or \code{"one.sided"}.
This controls whether the final combined \emph{p}-value is symmetrized.
\strong{Note:} To construct valid \emph{p}-value functions, the \code{\link{confMeta}}
function strictly requires \code{"two.sided"}.}
}
\value{
A numeric vector of combined \emph{p}-values corresponding
to each value of \code{mu}.
}
\description{
Stouffer's method for combining \emph{p}-values across studies
}
\details{
Stouffer's \emph{z}-statistic for \eqn{k} studies is defined as
\deqn{z = \frac{\sum_{i=1}^k w_i z_i}{\sqrt{\sum_{i=1}^k w_i^2}},}
where \eqn{i_i} and \eqn{{w_i}} are individual study \emph{z}-values and
weights, respectively. Under the global null hypothesis, each \eqn{z_i} is
assumed to follow a standard normal distribution. Stouffer's \emph{z}-statistic
then also follows a standard normal distribution. The combined \emph{p}-value
is calculated as the probability of observing a value equal or greater than
\eqn{z} from this distribution: \deqn{p_S = \Pr(Z > z)}
}
\examples{
estimates <- c(0.1, 0.2, 0.3)
SEs <- c(0.05, 0.05, 0.1)
p_stouffer(estimates, SEs, mu = 0, heterogeneity = "none")
}
\references{
Stouffer SA, et al. \emph{The American Soldier}. Princeton University Press, 1949.
\doi{10.2307/2572105}

Held, L, Hofmann, F, Pawel, S. (2025). A comparison of combined \emph{p}-value
functions for meta-analysis. \emph{Research Synthesis Methods}, 16:758-785.
\doi{10.1017/rsm.2025.26}
}
\seealso{
Other p-value combination functions: 
\code{\link{p_edgington}()},
\code{\link{p_edgington_w}()},
\code{\link{p_fisher}()},
\code{\link{p_hmean}()},
\code{\link{p_pearson}()},
\code{\link{p_tippett}()},
\code{\link{p_wilkinson}()}
}
\concept{p-value combination functions}
