% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfun_stouffer.R
\name{p_stouffer}
\alias{p_stouffer}
\title{Stouffer's method}
\usage{
p_stouffer(
  estimates,
  SEs,
  mu = 0,
  heterogeneity = "none",
  phi = NULL,
  tau2 = NULL,
  check_inputs = TRUE,
  alternative = "two.sided",
  w = NULL
)
}
\arguments{
\item{estimates}{Numeric vector of study-level effect estimates.}

\item{SEs}{Numeric vector of corresponding standard errors.}

\item{mu}{Numeric \strong{scalar or vector} of null values for the overall effect
(default: 0).}

\item{heterogeneity}{Character string: \code{"none"} (default),
\code{"additive"}, or \code{"multiplicative"}. Determines whether
standard errors are adjusted for between-study heterogeneity using
\code{tau2} or \code{phi}.}

\item{phi}{Multiplicative heterogeneity parameter (if applicable).}

\item{tau2}{Additive heterogeneity parameter (if applicable).}

\item{check_inputs}{Logical (default \code{TRUE}). If \code{TRUE},
perform input validation.}

\item{alternative}{???}

\item{w}{Numeric vector of weights. ???}
}
\value{
A numeric vector of combined \emph{p}-values corresponding
to each value of \code{mu}.
}
\description{
Stouffer's method for combining \emph{p}-values across studies
}
\details{
Explain how it is computed ???
}
\examples{
estimates <- c(0.1, 0.2, 0.3)
SEs <- c(0.05, 0.05, 0.1)
p_stouffer(estimates, SEs, mu = 0, heterogeneity = "none")
}
\references{
Stouffer SA, et al. \emph{The American Soldier}. Princeton University Press, 1949.

Held L, Hofmann F, Pawel S. A comparison of combined p-value functions for meta-analysis. \emph{Research Synthesis Methods}, 16:758-785, 2025.
}
\seealso{
Other p-value combination functions: 
\code{\link{p_edgington}()},
\code{\link{p_edgington_w}()},
\code{\link{p_fisher}()},
\code{\link{p_hmean}()},
\code{\link{p_pearson}()},
\code{\link{p_tippett}()},
\code{\link{p_wilkinson}()}
}
\concept{p-value combination functions}
